% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_heatmap.R
\name{annotation_heatmap}
\alias{annotation_heatmap}
\title{Annotation Heatmap}
\usage{
annotation_heatmap(
  effects_matrix,
  select_features = c("both", "distinctive", "largest"),
  feature_sign = c("both", "positive", "negative"),
  dims = colnames(effects_matrix),
  compare_dims = dims
)
}
\arguments{
\item{effects_matrix}{Describe the effects_matrix input argument here.}

\item{select_features}{Describe the select_features input argument here.}

\item{feature_sign}{Describe the feature_sign input argument here.}

\item{dims}{Describe the dims input argument here.}

\item{compare_dims}{Describe the compare_dims argument here.}
}
\value{
TO DO.
}
\description{
TO DO.
}
\examples{
# TO DO.

}
