% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{compute_binom_logbayesfactors}
\alias{compute_binom_logbayesfactors}
\title{Function Title Goes Here}
\usage{
compute_binom_logbayesfactors(X, fit, verbose = TRUE)
}
\arguments{
\item{X}{Describe input argument "X".}

\item{fit}{Describe input argument "fit",}

\item{verbose}{Describe input argument "verbose".}
}
\description{
Description of function goes here.
}
