% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{cost}
\alias{cost}
\title{Cost Function for Non-negative Matrix Factorization and
  Topic Modeling}
\usage{
cost(X, A, B, e = 1e-15, model = c("poisson", "multinom"), const = 0,
  version)
}
\arguments{
\item{X}{The n x m matrix of counts or pseudocounts. It can be a
sparse matrix ("dgCMatrix" class) or dense matrix ("matrix" class).}

\item{A}{The n x k matrix of loadings (also called "activations").
It should be a dense matrix; that is, \code{is.matrix(A)} should
be \code{TRUE}.}

\item{B}{The k x m matrix of factors (also called "basis vectors").
It should be a dense matrix; that is, \code{is.matrix(B)} should be
\code{TRUE}.}

\item{e}{A small, non-negative number added to the terms inside the
logarithms to avoid computing logarithms of zero. This prevents
numerical problems at the cost of introducing a very small
inaccuracy in the computation.}

\item{model}{If \code{model = "poisson"}, Poisson log-likelihoods
are returned; if \code{model = "multinom"}, multinomial
log-likelihoods are returned. See "Value" for details.}

\item{const}{TO DO: Add description here.}

\item{version}{If \code{version == "R"}, the computations are
performed entirely in R; if \code{version == "Rcpp"}, an Rcpp
implementation is used.  The R version is typically faster, but the
Rcpp version makes much more efficient use of memory for large,
sparse matrices.}
}
\value{
The return value is a vector with one entry per row of
  X. When \code{poisson = TRUE}, this vector contains the negative
  Poisson log-likelihoods; when \code{poisson = FALSE}, this vector
  contains the negative multinomial log-likelihoods.
}
\description{
Compute negative log-likelihoods for assessing
  a topic model fit or quality of a non-negative matrix factorization,
  in which matrix X is approximated by matrix product A * B.

  This function is mainly used internally to quickly compute
  log-likelihoods for model fits and objective values, and should not
  be used directly unless you know what you are doing. In particular,
  very little argument checking is performed; if you use this
  function, it is up to you to make sure you use it correctly.
}
\examples{

}
\seealso{

}
\keyword{internal}
