% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcount.R
\name{diff_count_analysis}
\alias{diff_count_analysis}
\title{Add Title Here}
\usage{
diff_count_analysis(X, fit)
}
\arguments{
\item{X}{Describe input argument "X" here.}

\item{fit}{Describe input argument "fit" here.}
}
\value{
Describe the return value here.
}
\description{
Describe function here.
}
