% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet.R
\docType{data}
\name{droplet}
\alias{droplet}
\title{Droplet single-cell RNA-seq read count data from Montoro
  \emph{et al} (2018)}
\format{\code{droplet} is a 7,193 x 17,133 sparse matrix.
\emph{Add more details about data format here.}}
\description{
Add details about data set here.
}
\examples{

# Describe example here.
library(Matrix)
data(droplet)
# Add example here.

}
\references{
D. T. Montoro emph{et al} (2018). A revised airway epithelial
hierarchy includes CFTR-expressing ionocytes. \emph{Nature} \bold{560},
319–-324.
}
\keyword{data}
