% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_plots.R
\name{embedding_plot_2d}
\alias{embedding_plot_2d}
\alias{embedding_plot_2d_ggplot_call}
\alias{pca_plot}
\alias{pca_hexbin_plot}
\alias{pca_hexbin_plot_ggplot_call}
\alias{tsne_plot}
\title{PCA Plot}
\usage{
embedding_plot_2d(
  fit,
  Y,
  fill = "loading",
  k,
  fill.label,
  ggplot_call = embedding_plot_2d_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots)
)

embedding_plot_2d_ggplot_call(
  Y,
  fill,
  fill.type = c("loading", "numeric", "factor", "none"),
  fill.label,
  font.size = 9
)

pca_plot(
  fit,
  k,
  out.pca,
  pcs = 1:2,
  fill = "loading",
  ggplot_call = pca_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)

pca_hexbin_plot(
  fit,
  out.pca,
  pcs = 1:2,
  bins = 40,
  breaks = c(0, 1, 10, 100, 1000, Inf),
  ggplot_call = pca_hexbin_plot_ggplot_call,
  ...
)

pca_hexbin_plot_ggplot_call(dat, pcs, bins, breaks, font.size = 9)

tsne_plot(
  fit,
  color = c("mixprop", "loading"),
  k,
  tsne,
  ggplot_call = tsne_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}.}

\item{Y}{Description of input argument "Y" goes here.}

\item{fill}{The quantity to map onto the fill colour of the points
in the PCA plot. Set \code{fill = "loading"} to vary the fill
colour according to the loadings (or topic proportions) of the
select topic, or topics. Alternatively, \code{fill} may be set to a
data vector with one entry per row of \code{fit$L}, in which case
the data are mapped to the fill colour of the points. When
\code{fill = "none"}, the fill colour is not varied.}

\item{k}{The topic, or topics, selected by number or name. One plot
is created per selected topic. When not specified, all topics are
plotted.}

\item{fill.label}{The label used for the fill colour legend.}

\item{ggplot_call}{The function used to create the plot. Replace
\code{tsne_plot_ggplot_call} with your own function to
customize the appearance of the plot.}

\item{plot_grid_call}{When multiple topics are selected, this is
the function used to arrange the plots into a grid using
\code{\link[cowplot]{plot_grid}}. It should be a function accepting
a single argument, \code{plots}, a list of \code{ggplot} objects.}

\item{fill.type}{The type of variable mapped to fill colour. The
fill colour is not varied when \code{fill.type = "none"}.}

\item{font.size}{Font size used in plot.}

\item{out.pca}{A list containing the result of a principal
components analysis, typically the result of calling
\code{\link[stats]{prcomp}} or a comparable function such as
\code{rpca} from the rsvd package. It should be a list object
containing, at a minimum, a list matrix or data frame
\code{out.pca$x} storing the rotated data. If not provided,
principal components will be computed automatically by calling
\code{\link[stats]{prcomp}}.}

\item{pcs}{The two principal components (PCs) to be plotted,
specified by name or number.}

\item{\dots}{Additional arguments passed to
\code{tsne_from_topics}. These arguments are only used if
\code{tsne} is not provided.}

\item{bins}{Number of bins used to create hexagonal 2-d
histogram. Passed as the \dQuote{bins} argument to
\code{\link[ggplot2]{stat_bin_hex}}.}

\item{breaks}{To produce the hexagonal histogram, the counts are
subdivided into intervals based on \code{breaks}. Passed as the
\dQuote{breaks} argument to \code{\link{cut}}.}

\item{color}{The data mapped to the color \dQuote{aesthetic} in the
plot. When \code{color = "loading"}, the estimated loadings (stored
the \code{L} matrix) in the Poisson NMF model are plotted; when
\code{color = "mixprop"}, the estimated mixture proportions (which are
recovered from the loadings by calling
\code{\link{poisson2multinom}} are shown. When \code{fit} is a
\dQuote{multinom_topic_model_fit} object, the only available option
is \code{color = "mixprop"}. In most settings, the mixture proportions
are preferred, even if the 2-d embedding is computed from the
loading matrix of the Poisson NMF model.}

\item{tsne}{A 2-d embedding of the samples (rows of X), or a subset
of the samples, such as an output from
\code{\link{tsne_from_topics}}. It should be a list object with the
same structure as a \code{tsne_from_topics} output; see
\code{\link{tsne_from_topics}} for details. If not provided, a 2-d
t-SNE embedding will be estimated automatically by calling
\code{tsne_from_topics}.}
}
\value{
A \code{ggplot} object.

A \code{ggplot} object.
}
\description{
Visualize the structure of the Poisson NMF loadings or
  the multinomial topic model mixture proportions by projection onto
  principal components (PCs). \code{plot_hexbin_plot} is most useful
  for visualizing the PCs of a data set with thousands of samples, or
  more.

Visualize the "structure" of the Poisson NMF loadings
  or multinomial topic model mixture proportions by projection onto a
  2-d surface. Samples in the projection are colored according to the
  their loadings/proportions. By default, t-SNE is used to compute
  the 2-d embedding from the loadings or mixture proportions.
}
\details{
This is a lightweight interface for rapidly producing PCA
plots from a Poisson non-negative matrix factorization or
multinomial topic model. The plotting is implemented using
ggplot2. For more control over the plot's appearance, the plot can
be customized by modifying the \code{ggplot_call} and
\code{plot_grid_call} arguments. If not provided, the principal
components are computed using \code{\link[stats]{prcomp}}.

Since principal components are a \emph{linear} projection
(\dQuote{rotation}) of the data, distances between points can bee
more interpretable than nonlinear embedding methods such as t-SNE
(as visualized using \code{tsne_plot}) and UMAP.

This is a lightweight interface primarily intended to
expedite creation of scatterplots for visualizing the loadings or
mixture proportions in 2-d; most of the \dQuote{heavy lifting} is
done by ggplot2. The 2-d embedding itself is computed by invoking
function \code{\link{tsne_from_topics}} (unless the \dQuote{tsne}
input is provided). For more control over the plot's appearance,
the plot can be customized by modifying the \code{ggplot_call} and
\code{plot_grid_call} arguments.

An effective 2-d visualization may also necessitate some
fine-tunning of the t-SNE settings, such as the
\dQuote{perplexity}, or the number of samples included in the
plot. The t-SNE settings can be controlled by the additional
arguments (\dots) passed to \code{tsne_from_topics}; see
\code{\link{tsne_from_topics}} for details. Alternatively, a 2-d
embedding may be pre-computed, and passed as argument \code{tsne}
to \code{tsne_plot}.
}
