% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{pca_from_topics}
\alias{pca_from_topics}
\alias{umap_from_topics}
\title{Low-dimensional Embeddings from Poisson NMF or Multinomial Topic Model}
\usage{
pca_from_topics(fit, dims = 2, center = TRUE, scale. = FALSE, ...)

umap_from_topics(fit, dims, ...)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{dims}{Describe input argument "dims" here.}

\item{center}{Describe input argument "center" here.}

\item{scale.}{Describe input argument "scale." here.}

\item{\dots}{Describe additional inputs here.}
}
\description{
Describe pca_from_topics here.
}
\details{
Add details here.
}
\seealso{
\code{\link{pca_plot}}, \code{\link{tsne_plot}},
  \code{\link{umap_plot}}
}
