% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lda.R
\name{fit_lda}
\alias{fit_lda}
\title{Add Title Here.}
\usage{
fit_lda(
  X,
  fit0,
  estimate.alpha = TRUE,
  iter.max = 20,
  minval = 1e-08,
  control = list(estimate.alpha = estimate.alpha, alpha = 1, verbose = 1, keep = 1, em
    = list(iter.max = iter.max, tol = 0), var = list(iter.max = 10, tol = 0)),
  ...
)
}
\arguments{
\item{X}{Describe input argument "X" here.}

\item{fit0}{Describe input argument "fit0" here.}

\item{estimate.alpha}{Describe input argument "estimate.alpha" here.}

\item{iter.max}{Describe input argument "numiter" here.}

\item{minval}{Describe input argument "minval" here.}

\item{control}{Describe input argument "control" here.}

\item{\dots}{Describe input arguments "..." here.}
}
\description{
Add description here.
}
\examples{
library(Matrix)
set.seed(1)
X <- simulate_count_data(80,100,k = 4,sparse = TRUE)$X
fit <- fit_topic_model(X,k = 4)
lda <- fit_lda(X,fit)

}
