% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multinom_model.R
\name{fit_multinom_model}
\alias{fit_multinom_model}
\title{Fit Simple Multinomial Model}
\usage{
fit_multinom_model(cluster, X, verbose = FALSE, ...)
}
\arguments{
\item{cluster}{A factor specifying a grouping, or clustering, of
the rows of \code{X}; e.g., the \dQuote{cluster} output from
\code{\link[stats]{kmeans}}.}

\item{X}{The n x m matrix of counts; all entries of X should be
non-negative. It can be a sparse matrix (class \code{"dgCMatrix"})
or dense matrix (class \code{"matrix"}), with some exceptions (see
\sQuote{Details}).}

\item{verbose}{This is passed as the \dQuote{verbose} argument in
the call to \code{\link{init_poisson_nmf}}.}

\item{\dots}{Additional arguments passed to
\code{\link{init_poisson_nmf}}.}
}
\value{
A multinomial topic model fit.
}
\description{
Give short description here.
}
\details{
The MLE has a closed-form solution; no iterative updates
  are needed.
}
\seealso{
\code{\link{fit_topic_model}}
}
