% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_topic_model.R
\name{fit_topic_model_simple}
\alias{fit_topic_model_simple}
\title{Simple Interface for Fitting a Multinomial Topic Model}
\usage{
fit_topic_model_simple(
  X,
  k,
  init.method = c("topicscore", "random"),
  control.init = list(),
  verbose = TRUE
)
}
\arguments{
\item{X}{The n x m matrix of counts; all entries of X should be
non-negative. It can be a sparse matrix (class \code{"dgCMatrix"})
or dense matrix (class \code{"matrix"}).}

\item{k}{The number of topics. Must be 2 or greater.}

\item{init.method}{The method used to initialize the factors and
loadings. See \code{\link{init_poisson_nmf}} for details.}

\item{control.init}{A list of parameters controlling the behaviour
of the optimization and Topic SCORE method in the call to
\code{init_poisson_nmf}.}

\item{verbose}{When \code{verbose = TRUE}, information about the
progress of the model fitting is printed to the console. See
\code{\link{fit_poisson_nmf}} for an explanation of the output.}
}
\value{
A multinomial topic model fit; see
  \code{\link{poisson2multinom}} and \code{\link{fit_poisson_nmf}}
  for details.
}
\description{
Fits a multinomial topic model to the count data,
  hiding most of the complexities of model fitting. The default
  optimization settings used here are intended to work well in a wide
  range of data sets, although some fine-tuning may be needed for
  more difficult cases.
}
\details{
TO DO: Add details here.
}
\examples{
library(Matrix)
set.seed(1)
X <- simulate_count_data(80,100,k = 3,sparse = TRUE)$X

}
\seealso{
\code{\link{init_poisson_nmf}},
  \code{\link{fit_poisson_nmf}}, \code{\link{poisson2multinom}}.
}
