% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{loadings_plot}
\alias{loadings_plot}
\title{Loadings Plot}
\usage{
loadings_plot(
  fit,
  k,
  x,
  geom.boxplot.params = list(width = 0.25, size = 0.4, outlier.shape = NA),
  labs.params = list(x = "", y = "loading"),
  use.theme = function() theme_cowplot(12) + theme(axis.line = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 1), plot.title = element_text(size =
    12, face = "plain"))
)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{k}{Describe input argument "k" here.}

\item{x}{Describe input argument "x" here.}

\item{geom.boxplot.params}{Describe input argument
"geom.boxplot.params" here.}

\item{labs.params}{Describe input argument "labs.params" here.}

\item{use.theme}{Describe input argument "use.theme" here.}
}
\description{
Add description here.
}
\details{
Add details here.
}
