% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_topics.R
\name{merge_topics}
\alias{merge_topics}
\title{Merge Topics in Multinomial Topic Model Fit}
\usage{
merge_topics(fit, k)
}
\arguments{
\item{fit}{A multinomial topic model fit.}

\item{k}{The names or numbers of the topics to be combined. Two or
more topics should be chosen.}
}
\description{
Combine two or more topics in a multinomial topic
  model fit.
}
\details{
Loadings (the topic proportions matrix, \code{L}) are
  combined by summation, and factors (the \code{F} matrix) are
  combined by averaging.
}
