% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson2multinom.R
\name{multinom2poisson}
\alias{multinom2poisson}
\title{Get Poisson Non-Negative Matrix Factorization from Multinomial
  Topic Model}
\usage{
multinom2poisson(fit, X)
}
\arguments{
\item{fit}{Describe the "fit" argument here.}

\item{X}{Describe the "X" argument here.}
}
\value{
Describe the return value here.
}
\description{
This function recovers parameter estimates of the
  Poisson non-negative matrix factorization given parameter estimates
  for a multinomial topic model.
}
