% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pca_plot}
\alias{pca_plot}
\title{Title Goes Here.}
\usage{
pca_plot(
  fit,
  k,
  out.pca,
  pcs = 1:2,
  ggplot_call = tsne_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{k}{Describe input argument "k" here.}

\item{out.pca}{Describe input argument "out.pca" here.}

\item{pcs}{Describe input argument "pcs" here.}

\item{ggplot_call}{Describe input argument "k" here.}

\item{plot_grid_call}{Describe input argument "plot_grid_call" here.}
}
\value{
Describe return value here.
}
\description{
Description of function goes here.
}
