% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_loglik_vs_rank}
\alias{plot_loglik_vs_rank}
\title{Plot Log-Likelihood vs. Rank}
\usage{
plot_loglik_vs_rank(fits)
}
\arguments{
\item{fits}{Describe input argument "fits" here.}
}
\value{
Describe return value here.
}
\description{
Describe function here.
}
