% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_progress_poisson_nmf}
\alias{plot_progress_poisson_nmf}
\title{Add title here.}
\usage{
plot_progress_poisson_nmf(fits, color, linetype, size, e = 0.01)
}
\arguments{
\item{fits}{Describe "fits" input argument here.}

\item{color}{Describe "color" input argument here.}

\item{linetype}{Describe "linetype" input argument here.}

\item{e}{Describe "e" input argument here.}
}
\value{
Describe return value here.
}
\description{
Add description here.
}
\seealso{
\code{\link{fit_poisson_nmf}}
}
