% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnmfem.R
\name{pnmfem}
\alias{pnmfem}
\title{#' @title Title goes here.}
\usage{
pnmfem(X, L0, F0, numiter = 1000, minval = 1e-15, e = 1e-15,
  verbose = TRUE)
}
\arguments{
\item{X}{The n x m matrix of counts; all entries of X should be
non-negative. Both dense and sparse matrices (the "matrix" and
"dgCMatrix" classes) are supported.}

\item{L0}{This is the initial estimate of the loadings (also called
"activations"). It should an n x k matrix, where n is the number of
rows of X, and k > 1 is the rank of the matrix factorization. All
entries of L should be non-negative.}

\item{F0}{This is the initial estimate of the factors (also called
"basis vectors"). It should be an m x k matrix, where m is the
number of columns of X, and k > 1 is the rank of the matrix
factorization, or the number of topics. All entries of F should be
non-negative.}

\item{numiter}{The number of EM updates to perform.}

\item{minval}{A small positive constant used to safeguard the EM
updates. The EM updates are implemented as \code{F <-
pmax(F1,minval)} and \code{L <- pmax(L1,minval)}, where \code{F1}
and \code{L1} are the factors and loadings matrices obtained by
applying a single EM update. Setting \code{minval = 0} is allowed,
but the EM updates are not guaranteed to converge to a stationary
point without this safeguard, and a warning will be given in this
case.}

\item{e}{A small, non-negative number added to the terms inside the
logarithms to avoid computing logarithms of zero. This prevents
numerical problems at the cost of introducing a very small
inaccuracy in the computation.}

\item{verbose}{When \code{verbose = TRUE}, information about the
algorithm's progress is printed to the console at each iteration.}
}
\value{
\code{pnmfem} returns a list object with the following
elements:

\item{F}{A dense matrix containing estimates of the factors.}

\item{L}{A dense matrix containing estimates of the loadings.}

\item{progress}{A data frame containing more detailed information
  about the algorithm's progress. The data frame should have
  \code{numiter} rows. The columns of the data frame are: "iter", the
  iteration number; "loglik", the log-likelihood at the current
  factor and loading estimates; "dev", the deviance at the current
  factor and loading estimates; "delta.l", the largest change in the
  factors matrix; "delta.f", the largest change in the loadings
  matrix; and "timing", the elapsed time in seconds (based on
  \code{\link{system.time}}).}
}
\description{
Provide description here.
}
\details{
Provide more details go here.
}
\examples{

# TO DO: Add example here.

}
\references{
TO DO: Add reference(s).
}
\seealso{
\code{\link{betanmf}}
}
