% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{poisson2binom}
\alias{poisson2binom}
\title{Function Title Goes Here}
\usage{
poisson2binom(X, fit, e = 0.01, version = c("Rcpp", "R"))
}
\arguments{
\item{X}{Describe input argument "X".}

\item{fit}{Describe input argument "fit".}

\item{e}{Describe input argument "e".}

\item{version}{Describe input argument "version".}
}
\description{
Description of function goes here.
}
