% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer.R
\name{run_homer}
\alias{run_homer}
\title{Perform HOMER Motif Enrichment Analysis Using DE Positions}
\usage{
run_homer(
  de,
  k,
  positions,
  genome = "hg19",
  subset = function(postmean, lpval, lfsr, rank, quantile) lfsr < 0.05,
  homer.exec = "findMotifsGenome.pl",
  out.dir = tempdir(),
  homer.options = "-len 8,10,12 -size 200 -mis 2 -h -S 25 -p 1",
  verbose = TRUE
)
}
\arguments{
\item{genome}{Genome assembly; e.g., "hg19"}

\item{out.dir}{Output directory.}

\item{regions.file}{BED file of the peak regions.}

\item{homer.path}{Directory path to 'findMotifsGenome.pl' excutable file.}

\item{region.size}{Size of the regions used for motif finding (default=200)}

\item{motif.length}{Length of motifs to be found (default=8,10,12).
HOMER will find motifs of each size separately and then combine the results at the end.
The length of time it takes to find motifs increases greatly with increasing size.}

\item{optimize.count}{Number of mismatches allowed in global optimization phase (default=2)}

\item{n.motifs}{Number of motifs to find (default=25)}

\item{use.hypergeometric}{Whether to use hypergeometric distribution to score motif enrichment
By default, findMotifsGenome.pl uses the binomial distribution to score motifs, which is faster.
This works well when the number of background sequences greatly out number the target sequences.
If the number of background sequences is smaller than target sequences, it is recommended to use the hypergeometric distribution.}

\item{background}{BED file of the background regions}
}
\value{
A data frame of motif enrichment results loaed from HOMER's output 'knownResults.txt'
}
\description{
http://homer.ucsd.edu/homer/
See http://homer.ucsd.edu/homer/ngs/peakMotifs.html for more details.
}
