% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select.poisson_nmf_fit}
\alias{select.poisson_nmf_fit}
\title{Extract or Re-order Poisson NMF Factors and Loadings}
\usage{
\method{select}{poisson_nmf_fit}(.data, factors, loadings, ...)
}
\arguments{
\item{.data}{Poisson NMF fit; that is, an object of class
\dQuote{poisson_nmf_fit}, such as an output from
\code{fit_poisson_nmf}.}

\item{factors}{Describe "factors" here.}

\item{loadings}{Describe "loadings" here.}

\item{...}{Other arguments passed to the generic select function.}
}
\value{
A "
}
\description{
This function can be used to extract estimates for a
  subset of the count data, or to re-order the factors and loadings
  (corresponding to columns and rows of the data matrix,
  respectively).
}
\details{
Evaluation metrics such as the log-likelihood and deviance
  are invariant to re-ordering of the factors and loadings (rows and
  columns of the counts matrix), but extracting a subset of the
  factors and/or loadings will invalidate these metrics.
}
\seealso{
\code{\link{fit_poisson_nmf}}
}
