% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.multinom_topic_model_fit}
\alias{select.multinom_topic_model_fit}
\alias{select}
\alias{select_loadings}
\title{Extract or Re-order Data Rows in Multinomial Topic Model}
\usage{
\method{select}{multinom_topic_model_fit}(.data, loadings, ...)

select_loadings(.data, loadings, ...)
}
\arguments{
\item{.data}{Multinomial Topic Model fit; that is, an object of
class \dQuote{multinom_topic_model_fit}, such as an output from
\code{fit_topic_model}.}

\item{loadings}{Indices (names or numbers) giving data rows to
keep. If not specified, all rows are kept.}

\item{\dots}{Other arguments passed to the generic select function.}
}
\value{
A multinomial topic model fit containing the selected
  data rows only.
}
\description{
This function can be used to extract estimates for a
  subset of the count data, or to re-order the rows of the loadings
  matrix in a multinomial topic model fit.
}
