% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{simulate_count_data}
\alias{simulate_count_data}
\title{Simulate count data}
\usage{
simulate_count_data(n, m, fmax = 1, lmax = 1, sparse = FALSE)
}
\arguments{
\item{n}{Number of rows in simulated count matrix.}

\item{m}{Number of columns in simulated count matrix.}

\item{fmax}{Factors are drawn uniformly at random between 0 and
\code{fmax}.}

\item{lmax}{Loadings are drawn uniformly at random between 0 and
\code{lmax}.}

\item{sparse}{If \code{sparse = TRUE}, convert the counts matrix to
a sparse matrix in compressed, column-oriented format; see
\code{\link[Matrix]{sparseMatrix}}.}

\item{k}{Number of factors used to determine Poisson rates.}
}
\value{

}
\description{
Simulate an n x m matrix X such that X[i,j] is Poisson
with rate (mean) Y[i,j], such that Y = tcrossprod(L,F), L is an n x
k matrix, and F is an m x k matrix. The entries of matrix L are
drawn uniformly at random between 0 and lmax, and the entries of
matrix F are drawn uniformly at random between 0 and fmax. The
return value is a list containing the data matrix X, and the
factors used to simulate the data, L and F. This function is mainly
used internally to simulate small data sets for testing the package
functions.

Only minimal argument checking.
}
