% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsne_from_topics}
\alias{tsne_from_topics}
\title{t-SNE from Poisson NMF or Topic Model}
\usage{
tsne_from_topics(
  fit,
  dims = 2,
  pca = FALSE,
  normalize = FALSE,
  perplexity = 100,
  theta = 0.1,
  max_iter = 1000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}.}

\item{dims}{The number of dimensions in the t-SNE embedding; passed
as argument \dQuote{dims} to \code{\link[Rtsne]{Rtsne}}.}

\item{pca}{Whether to perform a PCA processing stepe in t-SNE;
passed as argument \dQuote{pca} to \code{\link[Rtsne]{Rtsne}}.}

\item{normalize}{Whether to normalize the data prior to running
t-SNE; passed as argument \dQuote{normalize} to
\code{\link[Rtsne]{Rtsne}}.}

\item{max_iter}{Maximum number of t-SNE iterations; passed as
argument \dQuote{max_iter} to \code{\link[Rtsne]{Rtsne}}.}

\item{verbose}{Describe "verbose" here.}

\item{...}{Additional arguments passed to
\code{\link[Rtsne]{Rtsne}}.}
}
\value{
Describe the return value here.
}
\description{
Computes a low-dimensional embededding of the data from
  the estimated loadings, or topic probabilities, using the t-SNE
  method.
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
