% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsne_plot}
\alias{tsne_plot}
\title{t-SNE Plot}
\usage{
tsne_plot(
  fit,
  k,
  Y,
  ggplot_call = tsne_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{k}{Describe input argument "k" here.}

\item{Y}{Describe input argument "Y" here.}

\item{\dots}{Describe other inputs ("...") here.}
}
\value{
Describe the return value here.
}
\description{
Describe function here.
}
