% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altsqp.R, R/misc.R
\name{altsqp}
\alias{altsqp}
\alias{altsqp_control_default}
\alias{loglik.poisson}
\alias{loglik.multinom}
\alias{poisson2multinom}
\title{Alternating SQP method for non-negative matrix factorization
  and topic modeling.}
\usage{
altsqp(X, F, L, numiter = 100, control = list(), verbose = TRUE)

altsqp_control_default()

loglik.poisson(fit)

loglik.multinom(X, F, L, e = .Machine$double.eps)

poisson2multinom(F, L)
}
\arguments{
\item{X}{Describe X here.}

\item{F}{Describe F here.}

\item{L}{Describe L here.}

\item{numiter}{Describe numiter here.}

\item{control}{Describe control here.}

\item{verbose}{Describe verbose here.}
}
\description{
Compute maximum-likelihood estimates for the Poisson
  topic model; equivalently, find a non-negative matrix factorization
  X = L*F' that optimizes the beta divergence objective.
}
