% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_heatmap.R
\name{annotation_heatmap}
\alias{annotation_heatmap}
\title{Annotation Heatmap}
\usage{
annotation_heatmap(
  effects_matrix,
  select_features = c("both", "distinctive", "largest", "all"),
  feature_sign = c("both", "positive", "negative"),
  dims = colnames(effects_matrix),
  compare_dims = colnames(effects_matrix),
  n = 2,
  show_dims = colnames(effects_matrix),
  zero_value = 0.01,
  font_size = 10
)
}
\arguments{
\item{effects_matrix}{Describe the effects_matrix input argument here.}

\item{select_features}{Describe the select_features input argument here.}

\item{feature_sign}{Describe the feature_sign input argument here.}

\item{dims}{Describe the dims input argument here.}

\item{compare_dims}{Describe the compare_dims argument here.}

\item{n}{Describe argument n here.}

\item{show_dims}{Describe argument show_dims here.}

\item{zero_value}{Describe the zero_value argument here.}

\item{font_size}{Describe the font_size argument here.}
}
\value{
A \code{ggplot} object.
}
\description{
TO DO.
}
\examples{
# TO DO.

}
