% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{de_analysis}
\alias{de_analysis}
\title{Differential Expression Analysis using a Topic Model}
\usage{
de_analysis(
  fit,
  X,
  s = rowSums(X),
  pseudocount = 0.01,
  fit.method = c("scd", "em", "mu", "ccd", "glm"),
  shrink.method = c("ash", "none"),
  lfc.stat = "vsnull",
  numiter = 20,
  minval = 1e-10,
  tol = 1e-08,
  conf.level = 0.9,
  ns = 1000,
  rw = 0.3,
  eps = 1e-15,
  nc = 1,
  verbose = TRUE
)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If a Poisson NMF fit is provided
as input, the corresponding multinomial topic model fit is
automatically recovered using \code{\link{poisson2multinom}}.}

\item{X}{The n x m counts matrix. It can be a sparse matrix (class
\code{"dgCMatrix"}) or dense matrix (class \code{"matrix"}).}

\item{s}{A numeric vector of length n determining how the rates are
scaled in the Poisson model. See \dQuote{Details} for guidance on
the choice of \code{s}.}

\item{pseudocount}{Observations with this value are added to the
counts matrix to stabilize maximum-likelihood estimation.}

\item{fit.method}{Method used to fit the Poisson models.  Note that
\code{fit.method = "glm"} is the slowest method, and is mainly used
for testing.}

\item{shrink.method}{Method used to stabilize the LFC estimates.
When \code{shrink.method = "ash"}, the "adaptive shrinkage" method
implemented in the ashr package is used. When \code{shrink.method =
"none"}, no stabilization is performed, and the \dQuote{raw} LFC
estimates are returned.}

\item{lfc.stat}{Description of "lfc.stat" input argument goes here.}

\item{numiter}{Maximum number of iterations performed in fitting
the Poisson models. When \code{fit.method = "glm"}, this is passed
as argument \code{maxit} to the \code{glm} function.}

\item{minval}{A small, positive number. All topic proportions less
than this value and greater than \code{1-minval} are set to this
value.}

\item{tol}{Controls the convergence tolerance for fitting the
Poisson models. When \code{fit.method = "glm"}, this is passed as
argument \code{epsilon} to function \code{glm}.}

\item{conf.level}{Describe input argument "conf.level" here.}

\item{ns}{Describe input argument "ns" here.}

\item{rw}{Describe input argument "rw" here.}

\item{eps}{A small, non-negative number added to the terms inside
the logarithms to avoid computing logarithms of zero.}

\item{nc}{Number of threads used in the multithreaded computations.}

\item{verbose}{When \code{verbose = TRUE}, progress information is
printed to the console.}
}
\value{
The return value is a list of m x k matrices, where m is
  the number of columns in the counts matrix, and k is the number of
  topics (for \code{de_clusters}, m is the number of
  clusters), and an additional vector:

\item{colmeans}{A vector of length m containing the count averages
  (\code{colMeans(X)}).}

\item{F0}{Estimates of the Poisson model parameters \eqn{f_0}.}

\item{F1}{Estimates of the Poisson model parameters \eqn{f_1}.}

\item{beta}{LFC estimates \code{beta = log2(F1/F0)}.}

\item{se}{Standard errors of the Poisson glm parameters \eqn{b =
  f_1 - f_0}. }

\item{Z}{z-scores for the Poisson glm parameters \eqn{b = f_1 -
  f_0}.}

\item{pval}{-log10 two-tailed p-values computed from the
  z-scores. In some extreme cases the calculations may produce zero
  or negative standard errors, in which case the standard errors are
  set to \code{NA} and the z-scores and -log10 p-values are set to
  zero.}
}
\description{
Implements methods for differential expression
  analysis using a topic model. These methods are motivated by gene
  expression studies, but could have other uses, such as identifying
  topic keywords.
}
\references{
Stephens, M. (2016). False discovery rates: a new deal.
\emph{Biostatistics} \bold{18}(2), kxw041.
\url{https://doi.org/10.1093/biostatistics/kxw041}

Zhu, A., Ibrahim, J. G. and Love, M. I. (2019). Heavy-tailed prior
distributions for sequence count data: removing the noise and
preserving large differences. \emph{Bioinformatics} \bold{35}(12),
2084–2092.
}
