% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcount.R
\name{diff_count_analysis}
\alias{diff_count_analysis}
\title{Differential Count Analysis with a Multinomial Topic Model}
\usage{
diff_count_analysis(
  fit,
  X,
  s = rowSums(X),
  numiter = 100,
  e = 1e-15,
  verbose = TRUE
)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If a Poisson NMF fit is provided
as input, the corresponding multinomial topic model fit is
automatically recovered using \code{link{poisson2multinom}}.}

\item{X}{The n x m counts matrix. It can be a sparse matrix (class
\code{"dgCMatrix"}) or dense matrix (class \code{"matrix"}).}

\item{s}{A numeric vector of length n determining how the rates are
scaled in the Poisson model. See \dQuote{Details} for guidance on
the choice of \code{s}.}

\item{numiter}{The number of EM updates performed to compute the
maximum-likelihood estimates of the Poisson model parameters.}

\item{e}{A small, positive scalar included in some computations to
avoid logarithms of zero and division by zero.}

\item{verbose}{When \code{verbose = TRUE}, progress information is
printed to the console.}
}
\value{
Describe the return value here.
}
\description{
Describe function here.
}
\details{
Add details here.
}
