% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcount.R
\name{diff_count_analysis}
\alias{diff_count_analysis}
\title{Differential Count Analysis with a Multinomial Topic Model}
\usage{
diff_count_analysis(
  fit,
  X,
  s = rowSums(X),
  numiter = 100,
  tol = 1e-08,
  e = 1e-15,
  betamax = 10,
  verbose = TRUE
)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If a Poisson NMF fit is provided
as input, the corresponding multinomial topic model fit is
automatically recovered using \code{\link{poisson2multinom}}.}

\item{X}{The n x m counts matrix. It can be a sparse matrix (class
\code{"dgCMatrix"}) or dense matrix (class \code{"matrix"}).}

\item{s}{A numeric vector of length n determining how the rates are
scaled in the Poisson model. See \dQuote{Details} for guidance on
the choice of \code{s}.}

\item{numiter}{The number of EM updates performed to compute the
maximum-likelihood estimates of the Poisson model parameters.}

\item{tol}{When \code{tol > 0}, the EM algorithm for computing
maximum-likelihood estimates will be stopped early when the largest
change between two successive updates is less than \code{tol}.}

\item{e}{A small, positive scalar included in some computations to
avoid logarithms of zero and division by zero.}

\item{betamax}{Since the calculation of the standard deviations and
z-scores can be numerically unstable for large log-fold change
estimates, particularly in large data sets, an upper limit on the
estimated log-fold change can be used to improve the numerical
accuracy of the calculations.}

\item{verbose}{When \code{verbose = TRUE}, progress information is
printed to the console.}
}
\value{
The return value is a list with six m x k matrices, where m
is the number of columns in the counts matrix, and k is the number of
topics, and an additional vector:

\item{colmeans}{A vector of length m containing the count averages
  (\code{colMeans(X)}).}

\item{F0}{Estimates of the Poisson model parameters \eqn{f_0}.}

\item{F1}{Estimates of the Poisson model parameters \eqn{f_1}.}

\item{beta}{Log-fold change estimates, \code{beta = log2(F1/F0)}.}

\item{se}{Standard errors for the log-fold change estimates.}

\item{Z}{Log-fold change z-scores.}

\item{pval}{-log10 two-tailed p-values computed from the z-scores.}
}
\description{
Implements methods for analysis of differential count
expression using a topic model. These methods are motivated by gene
expression studies, but could have other uses, such as identifying
\dQuote{key words} in topics derived from text documents. A special
case of "hard" topic assignments is also implemented---that is, the
topic proportions are all zeros and ones---which involves greatly
simplified (and faster) calculations.
}
\details{
The methods are based on the following univariate
(\dQuote{single-count}) Poisson model: \deqn{x_i ~ Poisson(s_i
\lambda_i),} in which the Poisson rates are defined as
\deqn{\lambda_i = (1 - q_i) f_0 + q_i f_1,} and where the two
unknowns to be estimated are \eqn{f_0, f_1 > 0.} This model is
applied separately to each count (column of the counts matrix), and
to each topic k. The \eqn{q_i}'s are the topic probabilities for a
given topic. An EM algorithm is used to compute maximum-likelihood
estimates (MLEs) of the two unknowns.

The log-fold change statistics are defined as the log-ratio of the
two Poisson rate parameters, \eqn{\beta = \log_2(f_1/f_0)}. This
statistic measures the increase (or decrease) in occurence in one
topic compared to all other topics. The use of the base-2 logarithm
comes from the convention used in gene expression studies.

When each \eqn{s_i} (specified by input argument \code{s}) is equal
the total count for sample i (this is the default setting in
\code{diff_count_analysis}), the Poisson model will closely
approximate a binomial model of the count data, so that the Poisson
model parameters \eqn{f_0, f_1} represent binomial
probabilities. In this case, \eqn{\beta} represents the log-fold
change in \emph{relative} occurrence. (This Poisson approximation
to the binomial is most accurate when the total counts
\code{rowSums(X)} are large and \eqn{f_0, f_1} are small.)

Other choices for \code{s} are possible, and implement different
normalization schemes for the counts, and different interpretations
of the log-fold change statistics. Setting \code{s} to all ones
implements the differential count analysis with no normalization,
and \eqn{\beta} represents the log-fold change in \emph{absolute}
occurrence. This choice could be appropriate in settings where the
total count is well-controlled across samples (rows of \code{X}).

The standard error and z-score calculations are based on a Laplace
approximation to the likelihood at the MLE. This is the same
strategy used to compute standard errors and z-scores in
\code{\link[stats]{glm}}.
}
