% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{fit_binom_topic_model}
\alias{fit_binom_topic_model}
\title{Function Title Goes Here}
\usage{
fit_binom_topic_model(
  X,
  fit,
  p0,
  p1,
  method = c("em", "optim"),
  e = 1e-15,
  numiter = 40,
  control = list(factr = 1e-14, maxit = 100),
  verbose = TRUE
)
}
\arguments{
\item{X}{Describe input argument X.}

\item{fit}{Describe input argument F.}

\item{p0}{Describe input argument "p0".}

\item{p1}{Describe input argument "p1".}

\item{method}{Describe input argument "method".}

\item{e}{Describe input argument "e".}

\item{numiter}{Describe input argument "numiter".}

\item{control}{Describe input argument "control".}

\item{verbose}{Describe input argument "verbose".}
}
\value{
Describe return value.
}
\description{
Description of function goes here.
}
\details{
Using this function requires some care; only minimal
argument checking is performed, and error messages may not be
helpful.
}
