% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{loglik_poisson_topic_model}
\alias{loglik_poisson_topic_model}
\alias{loglik_multinom_topic_model}
\title{Compute Topic Model Likelihoods}
\usage{
loglik_poisson_topic_model(X, fit, e = 1e-15)

loglik_multinom_topic_model(X, fit, e = 1e-15)
}
\description{
Compute Topic Model Likelihoods
}
