% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{loglik_poisson_topic_model}
\alias{loglik_poisson_topic_model}
\alias{loglik_multinom_topic_model}
\alias{deviance_poisson_topic_model}
\alias{cost}
\title{Topic Model Likelihoods and Deviances}
\usage{
loglik_poisson_topic_model(X, fit, e = 1e-15)

loglik_multinom_topic_model(X, fit, e = 1e-15)

deviance_poisson_topic_model(X, fit, e = 1e-15)

cost(X, A, B, e = 1e-15, family = c("poisson", "multinom"), version)
}
\arguments{
\item{X}{The n x m matrix of counts or pseudocounts. It can be a
sparse matrix (class \code{"dgCMatrix"}) or dense matrix (class
\code{"matrix"}).}

\item{fit}{A list containing two dense, non-negative matrices,
\code{fit$F} and \code{fit$L}. The former is an m x k matrix of
factors, and the latter is an n x k matrix of loadings.}

\item{e}{A small, non-negative number added to the terms inside the
logarithms to avoid computing logarithms of zero. This prevents
numerical problems at the cost of introducing a very small
inaccuracy in the computation.}

\item{A}{The n x k matrix of loadings. It should be a dense matrix.}

\item{B}{The k x m matrix of factors. It should be a dense matrix.}

\item{family}{If \code{model = "poisson"}, the loss function
corresponding ot the Poisson topic model is computed; if
\code{model = "multinom"}, multinomial loss function values are
returned. See "Value" for details.}

\item{version}{When \code{version == "R"}, the computations are
 performed entirely in R; when \code{version == "Rcpp"}, an Rcpp
implementation is used. The R version is typically faster when
 \code{X} is a dense matrix, whereas the Rcpp version is faster and
 more memory-efficient when \code{X} is a large, sparse matrix. When
 not specified, the most suitable version is called depending on
 whether \code{X} is dense or sparse.}
}
\value{
A numeric vector with one entry per row of \code{X}.
}
\description{
Compute log-likelihoods and deviances for assessing
  topic model fit or quality of a non-negative matrix factorization,
  in which matrix X is approximated by the low-rank matrix product
  L*F'.

  Function \code{cost} is mainly used internally to quickly compute
  log-likelihoods and deviances, and should not be used directly
  unless you know what you are doing. In particular, very little
  argument checking is performed.
}
\examples{

# Generate a small counts matrix.
set.seed(1)
out <- simulate_count_data(10,20,3)
X   <- out$X
fit <- out[c("F","L")]

# Compute the Poisson log-likelihoods and deviances.
data.frame(loglik   = loglik_poisson_topic_model(X,fit),
           deviance = deviance_poisson_topic_model(X,fit))

# Compute multinomial log-likelihoods.
loglik_multinom_topic_model(X,poisson2multinom(fit))

}
