% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{loadings_plot}
\alias{loadings_plot}
\alias{loadings_plot_ggplot_call}
\title{Loadings Plot}
\usage{
loadings_plot(
  fit,
  k,
  x,
  ggplot_call = loadings_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots)
)

loadings_plot_ggplot_call(dat, k, font_size = 10)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{k}{Describe input argument "k" here.}

\item{x}{Describe input argument "x" here.}

\item{ggplot_call}{Describe input argument "ggplot_call" here.}

\item{dat}{Describe "dat" input argument here.}

\item{font_size}{Describe "font_size" input argument here.}

\item{plot.title}{Describe "plot.title" input argument here.}
}
\value{
Describe the return value here.
}
\description{
Add description here.
}
\details{
Add details here.
}
