% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{loadings_plot}
\alias{loadings_plot}
\alias{loadings_plot_ggplot_call}
\title{Loadings Plot}
\usage{
loadings_plot(
  fit,
  k,
  x,
  ggplot_call = loadings_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots)
)

loadings_plot_ggplot_call(dat, topic.label, font.size = 10)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}.}

\item{k}{The topic, or topics, selected by number or name.}

\item{x}{A categorical variable represented as a
\code{\link{factor}}. It should have the same number of elements as
the number of rows in \code{fit$L}.}

\item{ggplot_call}{The function used to create the plot. Replace
\code{loadings_plot_ggplot_call} with your own function to
customize the appearance of the plot.}

\item{plot_grid_call}{When multiple topics are selected, this is
the function used to arrange the plots into a grid using
\code{\link[cowplot]{plot_grid}}. It should be a function accepting
a single argument, \code{plots}, a list of \code{ggplot} objects.}

\item{dat}{A data frame passed as input to
\code{\link[ggplot2]{ggplot}}, containing columns "x" and
"loading".}

\item{topic.label}{Describe "topic.label" input argument here. The
name or number of the topic being plotted; it is only used to
determine the plot title.}

\item{font.size}{Font size used in plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Generate one or more bar plots to visualize the
  relationship between the loadings, or topic probabilities, and a
  selected categorical variable (a factor).
}
