% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_topics.R
\name{merge_topics}
\alias{merge_topics}
\title{Combine Topics in Multinomial Topic Model}
\usage{
merge_topics(fit, k)
}
\arguments{
\item{fit}{A multinomial topic model fit.}

\item{k}{The names or numbers of the topics to be combined. Two or
more topics should be chosen.}
}
\description{
Combine two or more topics in a multinomial topic
  model fit.
}
\details{
Mixture proportions are combined by summation, and factors
  are combined by averaging.
}
