% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_loglik_vs_rank}
\alias{plot_loglik_vs_rank}
\title{Plot Log-Likelihood Versus Poisson NMF Rank}
\usage{
plot_loglik_vs_rank(fits)
}
\arguments{
\item{fits}{A list with two more list elements, in which each list
element is an object of class \code{"poisson_nmf_fit"}. If two or
more Poisson NMF fits shares the same rank, the largest
log-likelihood is plotted.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a plot showing the improvement in the
  log-likelihood as the rank of the matrix factorization or
  the number of topics ("k") increases.
}
