% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_loglik_vs_rank}
\alias{plot_loglik_vs_rank}
\alias{loglik_vs_rank_ggplot_call}
\title{Plot Log-Likelihood Versus Poisson NMF Rank}
\usage{
plot_loglik_vs_rank(fits, ggplot_call = loglik_vs_rank_ggplot_call)

loglik_vs_rank_ggplot_call(dat, font.size = 9)
}
\arguments{
\item{fits}{A list with two more list elements, in which each list
element is an object of class \code{"poisson_nmf_fit"}. If two or
more Poisson NMF fits shares the same rank, the largest
log-likelihood is plotted.}

\item{ggplot_call}{The function used to create the plot. Replace
\code{loglik_vs_rank_ggplot_call} with your own function to
customize the appearance of the plot.}

\item{dat}{A data frame passed as input to
\code{\link[ggplot2]{ggplot}}, containing, at a minimum, columns
\dQuote{x} and \dQuote{y}.}

\item{font.size}{Font size used in plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a plot showing the improvement in the
  log-likelihood as the rank of the matrix factorization or
  the number of topics (\dQuote{k}) increases.
}
