% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_progress_poisson_nmf}
\alias{plot_progress_poisson_nmf}
\title{Plot Progress of NMF Optimization Method Over Time}
\usage{
plot_progress_poisson_nmf(
  fits,
  y = c("loglik", "dev", "res"),
  add.point.every = 20,
  colors = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00",
    "#CC79A7"),
  linetypes = "solid",
  linesizes = 0.5,
  shapes = 21,
  fills = "white",
  e = 0.01,
  theme = function() theme_cowplot(font_size = 12)
)
}
\arguments{
\item{fits}{An object of class \code{"poisson_nmf_fit"}, or a
non-empty, named list in which each list element is an object of
class \code{"poisson_nmf_fit"}.}

\item{y}{Describe "y" input argument here.}

\item{add.point.every}{A positive integer giving the iteration
interval for drawing points on the progress curves. Set to
\code{Inf} to prevent points from being drawn on the plot.}

\item{colors}{Colours used to draw progress curves; passed as the
\code{values} input to \code{\link[ggplot2]{scale_color_manual}}.
If fewer colours than "fits" are given, the colours are recycled.}

\item{linetypes}{Line types used to draw progress curves; passed as
the \code{values} input to \code{\link[ggplot2]{scale_linetype_manual}}.
If fewer line types than "fits" are given, the line types are
recycled.}

\item{linesizes}{Line sizes used to draw progress curves; passed as
the \code{values} input to \code{\link[ggplot2]{scale_size_manual}}.
If fewer line sizes than "fits" are given, the line sizes are
 recycled.}

\item{shapes}{Shapes used to draw points at the selected
iterations; passed as the \code{values} input to
\code{\link[ggplot2]{scale_shape_manual}}. If fewer shapes than
"fits" are given, the shapes are recycled.}

\item{fills}{Fill colours used to draw points at the selected
iterations; passed as the \code{values} input to
\code{\link[ggplot2]{scale_fill_manual}}. If fewer fill colours
than "fits", are given, the fill colours are recycled.}

\item{e}{Describe "e" input argument here.}

\item{theme}{Describe "theme" input argument here.}
}
\value{
Describe return value here.
}
\description{
Add description here.
}
\seealso{
\code{\link{fit_poisson_nmf}}
}
