% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson2binom.R
\name{poisson2binom}
\alias{poisson2binom}
\title{Recover Binomial Topic Model Fit from Poisson NMF fit}
\usage{
poisson2binom(X, fit, numem = 20, umin = 0.0001, verbose = TRUE)
}
\arguments{
\item{X}{The n x m \dQuote{binary} matrix; all entries of X should
be between 0 and 1 (including 0 and 1). It can be a sparse matrix
(class \code{"dgCMatrix"}) or dense matrix (class \code{"matrix"}).}

\item{fit}{Describe input argument "fit" here.}

\item{numem}{Describe input argument "numem" here.}

\item{umin}{Describe input argument "umin" here.}

\item{verbose}{Describe input argument "verbose" here.}
}
\value{
Describe the output here.
}
\description{
Add a brief description
}
\details{
Describe the binomial topic model in more detail here.
}
\examples{
# See the vignette for an example.

}
