% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson2binom.R
\name{poisson2binom}
\alias{poisson2binom}
\title{Recover Binomial Topic Model Fit from Poisson NMF fit}
\usage{
poisson2binom(fit)
}
\description{
Recover Binomial Topic Model Fit from Poisson NMF fit
}
\examples{
# See the vignette for an example.

}
