% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson2binom.R
\name{poisson2binom}
\alias{poisson2binom}
\title{Recover Binomial Topic Model Fit from Poisson NMF fit}
\usage{
poisson2binom(X, fit, numiter = 20)
}
\arguments{
\item{X}{Describe input argument "X" here.}

\item{fit}{Describe input argument "fit" here.}
}
\value{
Describe the output here.
}
\description{
Add a brief description
}
\details{
Describe the binomial topic model in more detail here.
}
\examples{
# See the vignette for an example.

}
