% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{simulate_poisson_gene_data}
\alias{simulate_poisson_gene_data}
\title{Add Title Here.}
\usage{
simulate_poisson_gene_data(n, m, k)
}
\arguments{
\item{n}{Describe input argument "n" here.}

\item{m}{Describe input argument "m" here.}

\item{k}{Describe input argument "k" here.}
}
\description{
Add description here.
}
\details{
Add details here.

Note that only minimal argument checking is performed. This
function is mainly used for testing the implementation of the
differential count analysis
}
