% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{simulate_poisson_gene_data}
\alias{simulate_poisson_gene_data}
\title{Simulate Gene-Expression Data with Absolute Expression Changes}
\usage{
simulate_poisson_gene_data(n, m, k, s, sparse = FALSE)
}
\arguments{
\item{n}{Number of rows in the simulated count matrix. It should be
at least 2.}

\item{m}{Number of columns in the simulated count matrix. It should
be at least 2.}

\item{k}{Number of factors, or "topics", used to generate the
individual Poisson rates. The number of topics should be 2 or
greater.}

\item{s}{A vector of "size factors"; each row of the loadings
matrix \code{L} is scaled by the entries of \code{s} before
generating the counts. This should be a vector of length n
containing only positive values.}

\item{sparse}{If \code{sparse = TRUE}, convert the counts matrix to
a sparse matrix in compressed, column-oriented format; see
\code{\link[Matrix]{sparseMatrix}}.}
}
\value{
The return value is a list containing the counts matrix
\code{X} and size factors \code{s} and factorization, \code{F},
\code{L}, used to generate the counts.
}
\description{
Simulate count data from a Poisson NMF model, in which
  topics represent \dQuote{gene expression programs}, and the gene
  expression programs are characterized by \emph{absolute} changes in
  expression rates. (This contrasts with
  \code{\link{simulate_multinom_gene_data}}, where the gene programs
  are characterized by relative changes in expression levels.) The
  way in which the counts are simulated is modeled after gene
  expression studies in which expression is measured by single-cell
  RNA sequencing (\dQuote{RNA-seq}) techniques: each row of the
  counts matrix corresponds a gene expression profile, each column
  corresponds to a gene, and each matrix element is a \dQuote{read
  count}, or \dQuote{UMI count}, measuring expression level. Factors
  are simulated so as to capture realistic changes in gene expression
  across different cell types. See \dQuote{Details} for the procedure
  used to simulate factors, loadings and counts.
}
\details{
Here we describe the process for generating the n x k
  loadings matrix \code{L} and the m x k factors matrix \code{F}. A
  key feature of the simulation process is that factors can be
  interpreted as Poisson rates, so that individual counts can be
  thought of as being generated from a weighted mixture of "topics"
  with different gene expression rates.

  Each row of the \code{L} matrix is generated in the following
  manner: (1) the number of nonzero topic proportions is \eqn{1
  \le n \le k}, with probability proportional to \eqn{2^{-n}};
  (2) the indices of the nonzero topic proportions are sampled
  uniformly at random; and (3) the nonzero topic proportions are
  sampled from the Dirichlet distribution with \eqn{\alpha = 1} (so
  that all topics are equally likely).

  Each row of the factors matrix (also Poisson rates) are
  generated according to the following procedure: (1) generate \eqn{u
  = |r| - 5}, where \eqn{r ~ N(0,2)}; (2) for each topic \eqn{k},
  generate the Poisson rates as \eqn{exp(max(t,-5))}, where \eqn{t ~
   0.95 * N(u,s/10) + 0.05 * N(u,s)}, and \eqn{s = exp(-u/8)}.

  Once the loadings and factors have been generated, the
  counts are simulated from the Poisson NMF model; that is,
  \code{X[i,j]} is Poisson with rate \code{Y[i,j]}, where \code{Y =
  tcrossprod(L,F)}.

  Note that only minimal argument checking is performed;
  the function is mainly used to test implementation of the
  topic-model-based differential count analysis.
}
