% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{structure_plot}
\alias{structure_plot}
\title{Structure Plot}
\usage{
structure_plot(fit)
}
\arguments{
\item{fit}{Describe}
}
\description{
Describe the structure plot. Explain why it is called
a "structure plot" (add reference(s) below).
}
\references{
[Add reference(s) here.]
}
