% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{structure_plot}
\alias{structure_plot}
\alias{structure_plot_ggplot_call}
\title{Structure Plot}
\usage{
structure_plot(
  fit,
  n = 2000,
  rows,
  grouping,
  topics,
  colors = c("#8dd3c7", "#ffffb3", "#bebada", "#fb8072", "#80b1d3", "#fdb462",
    "#b3de69", "#fccde5", "#d9d9d9", "#bc80bd", "#ccebc5", "#ffed6f"),
  gap = 1,
  ggplot_call = structure_plot_ggplot_call,
  ...
)

structure_plot_ggplot_call(dat, colors, ticks = NULL, font.size = 8)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If it is the former, the
multinomial topic model fit is automatically recovered using
\code{link{poisson2multinom}}.}

\item{n}{Describe input argument "n" here. Ignored if "rows" is
provided.}

\item{rows}{Ordering of rows (samples) in plot. Generated
automatically from 1-d t-SNE if not provided.}

\item{grouping}{Describe input argument "grouping" here.}

\item{topics}{Top-to-bottom order of topic probabilities in
structure plot. First topic is shown on the top. Default is
\code{1:k}, where \code{k} is the number of topics.}

\item{colors}{Colors used to draw topics in structure
plot: \code{colors[1]} is the colour used to draw \code{topics[1]},
\code{colors[2]} is the colour used to draw \code{topics[2]}, and
so on. Default colours are from \url{https://colorbrewer2.org}
(qualitative data, "12-class Set3").}

\item{gap}{Describe input argument "gap" here.}

\item{ggplot_call}{Describe input argument "ggplot_call" here.}

\item{...}{Additional arguments passed to
\code{\link{tsne_from_topics}}.}

\item{dat}{A data frame passed as input to
\code{\link[ggplot2]{ggplot}}, containing, at a minimum, columns
"sample", "topic" and "prob": column "sample" contains the
positions of the samples (rows of the loadings matrix) along the
horizontal axis; column "topic" is a topic (corresponding to
columns of the loadings matrix); and column "prob" is the topic
probability for the given sample.}

\item{ticks}{The placement of the group labels along the horizontal
axis, and their names. For data that is not grouped, use
\code{ticks = NULL.}}

\item{font.size}{Font size used in plot.}
}
\description{
Describe the structure plot. Explain why it is called
a "structure plot" (add reference(s) below).
}
\references{
[Add reference(s) here.]
}
