% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{structure_plot}
\alias{structure_plot}
\alias{plot.poisson_nmf_fit}
\alias{plot.multinom_topic_model_fit}
\alias{structure_plot_ggplot_call}
\title{Structure Plot}
\usage{
structure_plot(
  fit,
  n = 2000,
  rows,
  grouping,
  topics,
  colors = c("#e41a1c", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33",
    "#a65628", "#f781bf", "#999999"),
  gap = 1,
  perplexity = 100,
  eta = 200,
  ggplot_call = structure_plot_ggplot_call,
  ...
)

\method{plot}{poisson_nmf_fit}(x, ...)

\method{plot}{multinom_topic_model_fit}(x, ...)

structure_plot_ggplot_call(dat, colors, ticks = NULL, font.size = 9)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If a Poisson NMF fit is provided
as input, the corresponding multinomial topic model fit is
automatically recovered using \code{\link{poisson2multinom}}.}

\item{n}{The maximum number of samples (rows of the loadings
matrix) to include in the plot. Including large numbers (e.g.,
thousands) of samples is not recommended because it dramatically
slows down the t-SNE computation, and typically there is little to
no benefit in including large number of samples in the plot due to
screen resolution limits. Ignored if \code{rows} is provided.}

\item{rows}{Ordering of rows (samples) in plot, after they have
been grouped. Generated automatically from 1-d t-SNE if not
provided.}

\item{grouping}{Optional categorical variable (factor) with one
entry for each row of the loadings matrix \code{fit$L} defining a
grouping of the samples (rows). The samples (rows) are arranged
along the horizontal axis according to this grouping, then within
each group according to \code{rows} or, if not provided, according
to the 1-d t-SNE embedding.}

\item{topics}{Top-to-bottom ordering of the topics in the structure
plot; topics[1] is shown on the top, topics[2] is shown next, and
so on. If the ordering of the topics is not specified, the topics
are automaticcally ordered so that the topics with the greatest
"mass" are at shown at the bottom of the plot.}

\item{colors}{Colors used to draw topics in Structure plot:
\code{colors[1]} is the colour used to draw \code{topics[1]},
\code{colors[2]} is the colour used to draw \code{topics[2]}, and
so on. The default colour setting is the from
\url{https://colorbrewer2.org} (qualitative data, "9-class Set1").}

\item{gap}{The horizontal spacing between groups. Ignored if
\code{grouping} is not provided.}

\item{perplexity}{t-SNE perplexity parameter, passed as argument
\dQuote{perplexity} to \code{\link{tsne_from_topics}}. When
code{grouping} is provided, \code{perplexity} may be a vector of
settings of length equal to \code{levels(grouping)}, which allows
for a different t-SNE perplexity setting for each group.}

\item{eta}{t-SNE learning rate parameter, passed as argument
\dQuote{eta} to \code{\link{tsne_from_topics}}. When
code{grouping} is provided, \code{eta} may be a vector of
settings of length equal to \code{levels(grouping)}, which allows
for a different t-SNE learning rate for each group.}

\item{ggplot_call}{The function used to create the plot. Replace
\code{structure_plot_ggplot_call} with your own function to
customize the appearance of the plot.}

\item{\dots}{Additional arguments passed to \code{structure_plot}
(for the \code{plot} method) or \code{\link{tsne_from_topics}} (for
function \code{structure_plot}).}

\item{x}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If a Poisson NMF fit is provided
as input, the corresponding multinomial topic model fit is
automatically recovered using \code{\link{poisson2multinom}}.}

\item{dat}{A data frame passed as input to
\code{\link[ggplot2]{ggplot}}, containing, at a minimum, columns
"sample", "topic" and "mixprop": column "sample" contains the
positions of the samples (rows of the loadings matrix) along the
horizontal axis; column "topic" is a topic (corresponding to
columns of the loadings matrix); and column "mixprop" is the mixture
proportion for the given sample.}

\item{ticks}{The placement of the group labels along the horizontal
axis, and their names. For data that is not grouped, use
\code{ticks = NULL.}}

\item{font.size}{Font size used in plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a "Structure plot" from a multinomial topic
  model fit. The Structure plot represents the estimated mixture
  proportions of each sample as a stacked barchart, with bars of
  different colors representing different topics. Consequently,
  samples that have similar mixture proportions have similar amounts
  of each color.
}
\details{
The name "Structure plot" comes from its widespread use in
  population genetics to visualize the results of the Structure
  software (Rosenberg \emph{et al}, 2002).

  For most uses of the Structure plot in population genetics, there
  is usually some grouping of the samples (e.g., assignment to
  pre-defined populations) that guides arrangement of the samples
  along the horizontal axis in the bar chart. In other applications,
  such as analysis of gene expression data, no pre-defined grouping
  exists. Therefore, a "smart" arrangement of the samples is, by
  default, generated automatically by performing a 1-d t-SNE
  embedding of the samples.

  Alternatively, a categorical variable---the grouping---may be
  provided, in which case the samples are arranged according to that
  grouping, then arranged within each group using t-SNE.
}
\references{
Dey, K. K., Hsiao, C. J. and Stephens, M. (2017). Visualizing the
structure of RNA-seq expression data using grade of membership
models. \emph{PLoS Genetics} \bold{13}, e1006599.

Rosenberg, N. A., Pritchard, J. K., Weber, J. L., Cann, H. M.,
Kidd, K. K., Zhivotovsky, L. A. and Feldman, M. W. (2002). Genetic
structure of human populations. \emph{Science} \bold{298},
2381–2385.
}
