% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_plot.R
\name{structure_plot}
\alias{structure_plot}
\alias{plot.poisson_nmf_fit}
\alias{plot.multinom_topic_model_fit}
\alias{structure_plot_ggplot_call}
\title{Structure Plot}
\usage{
structure_plot(
  fit,
  grouping,
  topics,
  rows,
  n = 2000,
  colors = c("#e41a1c", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33",
    "#a65628", "#f781bf", "#999999"),
  gap = 1,
  embed_method = function(fit, ...) if (nrow(fit$L) < 20) rnorm(nrow(fit$L)) else
    drop(umap_from_topics(fit, dims = 1, ...)),
  ggplot_call = structure_plot_ggplot_call,
  ...
)

\method{plot}{poisson_nmf_fit}(x, ...)

\method{plot}{multinom_topic_model_fit}(x, ...)

structure_plot_ggplot_call(dat, colors, ticks = NULL, font.size = 9)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If a Poisson NMF fit is provided
as input, the corresponding multinomial topic model fit is
automatically recovered using \code{\link{poisson2multinom}}.}

\item{grouping}{Optional categorical variable (factor) with one
entry for each row of the loadings matrix \code{fit$L} defining a
grouping of the samples (rows). The samples (rows) are arranged
along the horizontal axis according to this grouping, then within
each group according to \code{rows} or, if not provided, according
to the 1-d UMAP embedding. If \code{grouping} is not a factor, an
attempt is made to convert it to a factor using \code{as.factor}.}

\item{topics}{Top-to-bottom ordering of the topics in the structure
plot; \code{topics[1]} is shown on the top, \code{topics[2]} is
shown next, and so on. If the ordering of the topics is not
specified, the topics are automaticcally ordered so that the topics
with the greatest \dQuote{mass} are at shown at the bottom of the
plot. The topics may be specified by number or by name.}

\item{rows}{Ordering of the rows (samples) in the structure plot
after they have been grouped. This may include all the rows, or a
subset. If not provided, generated automatically from a 1-d UMAP
embedding, separately for each group. The rows may be specified by
number or by name.}

\item{n}{The maximum number of samples (rows of the loadings
matrix) to include in the plot. Typically there is little to no
benefit in including large number of samples in the Structure plot
due to screen resolution limits. Ignored if \code{rows} is
provided.}

\item{colors}{Colors used to draw topics in Structure plot. The
default colour setting is the from \url{https://colorbrewer2.org}
(qualitative data, \dQuote{9-class Set1}).}

\item{gap}{The horizontal spacing between groups. Ignored if
\code{grouping} is not provided.}

\item{embed_method}{Describe input argument "embed_method" here.}

\item{ggplot_call}{The function used to create the plot. Replace
\code{structure_plot_ggplot_call} with your own function to
customize the appearance of the plot.}

\item{\dots}{Additional arguments passed to \code{structure_plot}
(for the \code{plot} method) or \code{\link{tsne_from_topics}} (for
function \code{structure_plot}).}

\item{x}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If a Poisson NMF fit is provided
as input, the corresponding multinomial topic model fit is
automatically recovered using \code{\link{poisson2multinom}}.}

\item{dat}{A data frame passed as input to
\code{\link[ggplot2]{ggplot}}, containing, at a minimum, columns
\dQuote{sample}, \dQuote{topic} and \dQuote{prop}: the
\dQuote{sample} column contains the positions of the samples (rows
of the L matrix) along the horizontal axis; the \dQuote{topic}
column is a topic (a column of L); and the \dQuote{prop} column is
the topic proportion for the respective sample.}

\item{ticks}{The placement of the group labels along the horizontal
axis, and their names. For data that are not grouped, use
\code{ticks = NULL}.}

\item{font.size}{Font size used in plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a \dQuote{Structure plot} from a multinomial topic
  model fit. The Structure plot represents the estimated mixture
  proportions of each sample as a stacked barchart, with bars of
  different colors representing different topics. Consequently,
  samples that have similar mixture proportions have similar amounts
  of each color.
}
\details{
The name \dQuote{Structure plot} comes from its widespread
use in population genetics to visualize the results of the
Structure software (Rosenberg \emph{et al}, 2002).

For most uses of the Structure plot in population genetics, there
is usually some grouping of the samples (e.g., assignment to
pre-defined populations) that guides arrangement of the samples
along the horizontal axis in the bar chart. In other applications,
such as analysis of gene expression data, no pre-defined grouping
exists. Therefore, a \dQuote{smart} arrangement of the samples is,
by default, generated automatically by performing a 1-d t-SNE
embedding of the samples.

Alternatively, a categorical variable---the grouping---may be
provided, in which case the samples are arranged according to that
grouping, then arranged within each group using t-SNE.
}
\examples{
# Add example(s) here.

}
\references{
Dey, K. K., Hsiao, C. J. and Stephens, M. (2017). Visualizing the
structure of RNA-seq expression data using grade of membership
models. \emph{PLoS Genetics} \bold{13}, e1006599.

Rosenberg, N. A., Pritchard, J. K., Weber, J. L., Cann, H. M.,
Kidd, K. K., Zhivotovsky, L. A. and Feldman, M. W. (2002). Genetic
structure of human populations. \emph{Science} \bold{298},
2381–2385.
}
