% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{structure_plot}
\alias{structure_plot}
\alias{structure_plot_ggplot_call}
\title{Structure Plot}
\usage{
structure_plot(
  fit,
  n = 2000,
  rows,
  grouping,
  topics,
  colors = c("#8dd3c7", "#ffffb3", "#bebada", "#fb8072", "#80b1d3", "#fdb462",
    "#b3de69", "#fccde5", "#d9d9d9", "#bc80bd", "#ccebc5", "#ffed6f"),
  gap = 1,
  ggplot_call = structure_plot_ggplot_call,
  ...
)

structure_plot_ggplot_call(dat, colors, ticks = NULL, font.size = 8)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}. If it is the former, the
multinomial topic model fit is automatically recovered using
\code{link{poisson2multinom}}.}

\item{n}{Describe input argument "n" here. Ignored if "rows" is
provided.}

\item{rows}{Ordering of rows (samples) in plot. Generated
automatically from 1-d t-SNE if not provided.}

\item{grouping}{Describe input argument "grouping" here.}

\item{topics}{Top-to-bottom order of topic probabilities in
structure plot. First topic is shown on the top. Default is
\code{1:k}, where \code{k} is the number of topics.}

\item{colors}{Describe input argument "colors" here.}

\item{gap}{Describe input argument "gap" here.}

\item{ggplot_call}{Describe input argument "ggplot_call" here.}

\item{...}{Additional arguments passed to
\code{\link{tsne_from_topics}}.}

\item{dat}{Describe input argument "dat" here.}

\item{ticks}{Describe input argument "ticks" here.}

\item{font.size}{Describe input argument "font.size" here.}
}
\description{
Describe the structure plot. Explain why it is called
a "structure plot" (add reference(s) below).
}
\references{
[Add reference(s) here.]
}
