% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.poisson_nmf_fit}
\alias{summary.poisson_nmf_fit}
\alias{print.summary.poisson_nmf_fit}
\title{Summarize Poisson NMF Fit}
\usage{
\method{summary}{poisson_nmf_fit}(object, ...)

\method{print}{summary.poisson_nmf_fit}(x, ...)
}
\arguments{
\item{object}{Describe input argument "object" here.}

\item{\dots}{Additional arguments passed to the generic method.}

\item{x}{An object of class \dQuote{summary.poisson_nmf_fit},
usually a result of a call to \code{summary.poisson_nmf_fit}.}
}
\value{
The function \code{summary.poisson_nmf_fit} computes and
  returns a list of statistics summarizing the model fit.
}
\description{
Add description of function here.
}
