% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsne_from_topics}
\alias{tsne_from_topics}
\title{t-SNE from Poisson NMF or Topic Model}
\usage{
tsne_from_topics(
  fit,
  dims = 2,
  n = 5000,
  pca = FALSE,
  normalize = FALSE,
  perplexity = 100,
  theta = 0.1,
  max_iter = 1000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}.}

\item{dims}{The number of dimensions in the t-SNE embedding; passed
as argument \dQuote{dims} to \code{\link[Rtsne]{Rtsne}}.}

\item{n}{The maximum number of rows in the loadings matrix
\code{fit$L} to use; when the loadings matrix has more than
\code{n} rows, the t-SNE embedding is computed on a random
selection of \code{n} rows.}

\item{pca}{Whether to perform a PCA processing stepe in t-SNE;
passed as argument \dQuote{pca} to \code{\link[Rtsne]{Rtsne}}.}

\item{normalize}{Whether to normalize the data prior to running
t-SNE; passed as argument \dQuote{normalize} to
\code{\link[Rtsne]{Rtsne}}.}

\item{max_iter}{Maximum number of t-SNE iterations; passed as
argument \dQuote{max_iter} to \code{\link[Rtsne]{Rtsne}}.}

\item{verbose}{If \code{verbose = TRUE}, progress updates are
printed; passed as argument \dQuote{verbose} to
\code{\link[Rtsne]{Rtsne}}.}

\item{\dots}{Additional arguments passed to
\code{\link[Rtsne]{Rtsne}}.}
}
\value{
A list with two elements: \code{Y}, an n x d matrix
  containing the embedding \code{Y} returned by
  \code{\link[Rtsne]{Rtsne}}, where n is the number of rows of the
  loadings matrix, and \code{d = dims}; \code{rows}, the rows of the
  loadings matrix included in the t-SNE embedding.
}
\description{
Computes a low-dimensional embededding of the data from
  the estimated loadings, or topic probabilities, using the t-SNE
  nonlinear dimensionality reduction method.
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
