% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsne_from_topics}
\alias{tsne_from_topics}
\title{t-SNE from Poisson NMF or Multinomial Topic Model}
\usage{
tsne_from_topics(
  fit,
  dims = 2,
  n = 5000,
  scaling = NULL,
  pca = FALSE,
  normalize = FALSE,
  perplexity = 100,
  theta = 0.1,
  max_iter = 1000,
  eta = 200,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}.}

\item{dims}{The number of dimensions in the t-SNE embedding; passed
as argument \dQuote{dims} to \code{\link[Rtsne]{Rtsne}}.}

\item{n}{The maximum number of rows in the loadings matrix
\code{fit$L} to use; when the loadings matrix has more than
\code{n} rows, the t-SNE embedding is computed on a random
selection of \code{n} rows. An upper limit on the number of rows is
used because the runtime of \code{\link[Rtsne]{Rtsne}} increases
rapidly with the number of rows in the input matrix.}

\item{scaling}{A numeric vector of length equal to the number of
topics specifying a scaling of the columns of \code{fit$L}; this
re-scaling is performed prior to running t-SNE. The vector should
contain non-negative numbers only. A larger value will increase the
importance, or "weight", of the respective topic in computing the
embedding. When \code{scaling} is \code{NULL}, no re-scaling is
performed. Note that this scaling will have no effect if
\code{normalize = TRUE}.}

\item{pca}{Whether to perform a PCA processing stepe in t-SNE;
passed as argument \dQuote{pca} to \code{\link[Rtsne]{Rtsne}}.}

\item{normalize}{Whether to normalize the data prior to running
t-SNE; passed as argument \dQuote{normalize} to
\code{\link[Rtsne]{Rtsne}}.}

\item{perplexity}{t-SNE perplexity parameter; passed as
argument \dQuote{perplexity} to \code{\link[Rtsne]{Rtsne}}. Note
that the default setting may not work for smaller data sets.}

\item{theta}{t-SNE speed/accuracy trade-off parameter; passed as
argument \dQuote{theta} to \code{\link[Rtsne]{Rtsne}}.}

\item{max_iter}{Maximum number of t-SNE iterations; passed as
argument \dQuote{max_iter} to \code{\link[Rtsne]{Rtsne}}.}

\item{eta}{t-SNE learning rate parameter; passed as argument
\dQuote{eta} to \code{\link[Rtsne]{Rtsne}}.}

\item{verbose}{If \code{verbose = TRUE}, progress updates are
printed; passed as argument \dQuote{verbose} to
\code{\link[Rtsne]{Rtsne}}.}

\item{\dots}{Additional arguments passed to \code{\link[Rtsne]{Rtsne}}.}
}
\value{
A list with two list elements: \code{Y}, an n x d matrix
  containing the embedding \code{Y} returned by
  \code{\link[Rtsne]{Rtsne}}, where n is the number of rows of the
  loadings matrix, and \code{d = dims}; \code{rows}, the rows of the
  loadings matrix included in the t-SNE embedding.
}
\description{
Computes a low-dimensional embededding of the data from
  the estimated loadings or topic proportions using the t-SNE
  nonlinear dimensionality reduction method.
}
\details{
This is a lightweight interface for rapidly producing
  t-SNE embeddings from matrix factorizations or multinomial topic
  models; in particular, \code{tsne_from_topics} replaces the t-SNE
  defaults with settings that are more suitable for visualizing the
  structure of a matrix factorization or topic model (e.g., the PCA
  step in \code{Rtsne} is activated by default, but disabled in
  \code{tsne_from_topics}). See Kobak and Berens (2019) for guidance
  on choosing t-SNE settings such as the "perplexity" and learning
  rate (\code{eta}).
}
\references{
Kobak, D. and Berens, P. (2019). The art of using t-SNE for
single-cell transcriptomics. \emph{Nature Communications} \bold{10},
5416. \url{https://doi.org/10.1038/s41467-019-13056-x}
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
