% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsne_from_topics}
\alias{tsne_from_topics}
\title{t-SNE from Poisson NMF or Topic Model}
\usage{
tsne_from_topics(
  fit,
  dims = 2,
  pca = FALSE,
  normalize = FALSE,
  perplexity = 100,
  theta = 0.1,
  max_iter = 1000,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{dims}{Describe input argument "dims" here.}

\item{pca}{Describe input argument "pca" here.}

\item{normalize}{Describe input argument "normalize" here.}

\item{max_iter}{Describe input argument "max_iter" here.}

\item{verbose}{Describe input argument "verbose" here.}

\item{...}{Describe other input arguments here.}
}
\value{
Describe the return value here.
}
\description{
Computes a 2-dimensional embededding of the data from
  the estimated topic probabilities, or "loadings", using the t-SNE
  method.
}
