% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tsne_from_topics}
\alias{tsne_from_topics}
\title{Title Goes Here}
\usage{
tsne_from_topics(
  fit,
  dims = 2,
  pca = FALSE,
  normalize = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{dims}{Describe input argument "dims" here.}

\item{pca}{Describe input argument "pca" here.}

\item{normalize}{Describe input argument "normalize" here.}

\item{verbose}{Describe input argument "verbose" here.}

\item{...}{Describe other input arguments here.}
}
\value{
Describe the return value here.
}
\description{
Describe function here.
}
