% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsne_plot}
\alias{tsne_plot}
\alias{tsne_plot_ggplot_call}
\title{t-SNE Plot}
\usage{
tsne_plot(
  fit,
  color = c("probs", "loadings"),
  k,
  tsne,
  ggplot_call = tsne_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)

tsne_plot_ggplot_call(dat, topic.label, font.size = 10)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{k}{Describe input argument "k" here.}

\item{tsne}{Describe input argument "tsne" here.}

\item{ggplot_call}{Describe input argument "ggplot_call" here.}

\item{plot_grid_call}{Describe input argument "plot_grid_call" here.}

\item{\dots}{Describe other inputs ("...") here.}

\item{dat}{Describe "dat" here.}

\item{topic.label}{Describe "topic.label" input argument here.}

\item{font.size}{Describe "font.size" input argument here.}
}
\value{
Describe the return value here.
}
\description{
Describe function here.
}
