% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsne_plot}
\alias{tsne_plot}
\title{Title Goes Here}
\usage{
tsne_plot(
  fit,
  k,
  Y,
  geom.point.params = list(color = "white", stroke = 0.3, shape = 21),
  scale.fill.gradient2.params = list(low = "lightskyblue", mid = "gold", high =
    "orangered", midpoint = 0.5),
  labs.params = list(x = "tsne 1", y = "tsne 2", title = paste("topic", k)),
  use.theme = function() theme_cowplot(12) + theme(axis.line = element_blank(),
    plot.title = element_text(size = 12, face = "plain"))
)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{k}{Describe input argument "k" here.}

\item{Y}{Describe input argument "Y" here.}
}
\value{
Describe the return value here.
}
\description{
Describe function here.
}
