% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsne_plot}
\alias{tsne_plot}
\alias{tsne_plot_ggplot_call}
\title{t-SNE Plot}
\usage{
tsne_plot(
  fit,
  color = c("probability", "loading"),
  k,
  tsne,
  ggplot_call = tsne_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)

tsne_plot_ggplot_call(dat, topic.label, font.size = 10)
}
\arguments{
\item{fit}{An object of class \dQuote{poisson_nmf_fit} or
\dQuote{multinom_topic_model_fit}.}

\item{color}{The data mapped to the color \dQuote{aesthetic} in the
plot. When \code{color = "loading"}, the estimated loadings (stored
the \code{L} matrix) in the Poisson NMF model are plotted; when
\code{color = "probability"}, the estimated topic probabilities
(which are recovered from the loadings by calling
\code{\link{poisson2multinom}} are shown. When \code{fit} is a
\dQuote{multinom_topic_model_fit} object, the only available option
is \code{color = "probability"}. In most settings, the topic
probabilities are preferred, even if the 2-d embedding is computed
from the loading matrix of the Poisson NMF model.}

\item{k}{The topic, or topics, selected by number or name. One plot
is created per selected topic. When not specified, all topics are
plotted.}

\item{tsne}{A 2-d embedding of the samples (rows of X), or a subset
of the samples, such as an output from
\code{\link{tsne_from_topics}}. It should be a list object with the
same structure as a \code{tsne_from_topics} output; see
\code{\link{tsne_from_topics}} for details. If not provided, a 2-d
t-SNE embedding will be estimated automatically by calling
\code{tsne_from_topics}.}

\item{ggplot_call}{The function used to create the plot. Replace
\code{tsne_plot_ggplot_call} with your own function to
customize the appearance of the plot.}

\item{plot_grid_call}{When multiple topics are selected, this is
the function used to arrange the plots into a grid using
\code{\link[cowplot]{plot_grid}}. It should be a function accepting
a single argument, \code{plots}, a list of \code{ggplot} objects.}

\item{\dots}{Additional arguments passed to
\code{tsne_from_topics}. These arguments are only used if
\code{tsne} is not provided.}

\item{dat}{A data frame passed as input to
\code{\link[ggplot2]{ggplot}}, containing, at a minimum, columns
"d1", "d2" (the first and second dimensions in the 2-d embedding),
and "loading".}

\item{topic.label}{The name or number of the topic being plotted;
it is only used to determine the plot title.}

\item{font.size}{Font size used in plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize the "structure" of the Poisson NMF loadings
  ("activations") or the topic probabilities by projection onto a 2-d
  surface. Samples in the projection are colored according to the their
  loadings/probabilities. By default, t-SNE is used to compute the
  2-d embedding from the loadings or topic probabilities.
}
\details{
This is a lightweight interface primarily intended to
  expedite creation of scatterplots for visualizing the loadings or
  topic probabilities in 2-d; most of the "heavy lifting" is done by
  ggplot2 (specifically, function \code{\link[ggplot2]{geom_boxplot}}
  in the ggplot2 package). The 2-d embedding itself is computed by
  invoking function \code{\link{tsne_from_topics}} (unless the "tsne"
  input is provided). For more control over the plot's appearance,
  the plot can be customized by modifying the \code{ggplot_call} and
  \code{plot_grid_call} arguments.
}
