% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcano_plot}
\alias{volcano_plot}
\alias{volcano_plot_ggplot_call}
\alias{plot.topic_model_diff_count}
\title{Volcano Plot}
\usage{
volcano_plot(
  diff_count_result,
  k,
  labels,
  betamax = 10,
  label_above_quantile = 0.99,
  ggplot_call = volcano_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots)
)

volcano_plot_ggplot_call(dat, topic.label, font.size = 9)

\method{plot}{topic_model_diff_count}(x, ...)
}
\arguments{
\item{diff_count_result}{An object of class
\dQuote{topic_model_diff_count}, usually an output from
\code{\link{diff_count_analysis}}.}

\item{k}{The topic, or topics, selected by number or name. When not
specified, all topics are plotted.}

\item{betamax}{Describe input argument "betamax" here.}

\item{label_above_quantile}{Describe inputargument
"label_above_quantile" here.}

\item{plot_grid_call}{When multiple topics are selected, this is
the function used to arrange the plots into a grid using
\code{\link[cowplot]{plot_grid}}. It should be a function accepting
a single argument, \code{plots}, a list of \code{ggplot} objects.}

\item{dat}{Describe input argument "dat" here.}

\item{topic.label}{The name or number of the topic being plotted.
Only used to determine the plot title.}

\item{font.size}{Describe input argument "font.size" here.}

\item{x}{Describe input argument "x" here.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a "volcano-like" plot, in which relative gene
  expression (x-axis), as measured by the log-fold change (lfc), is
  plotted against the z-score, or some other measure of support
  (y-axis). Points above a specified z-score quantile are labeled by
  the gene symbol. Since this scatterplot can have very many points
  (one for every gene), optional arguments "cutoff" and "n0" can be
  adjusted to control the number of points shown. The "cutoff"
  argument is a number between 0 and 1 controlling the quantile along
  the Y axis at which the points are thinned; the "n0" controls the
  number of points to randomly sample from the "thinned" part of the
  plot.
}
\details{
Add details here.
}
