% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plots.R
\name{volcano_plot}
\alias{volcano_plot}
\alias{plot.topic_model_de_analysis}
\alias{volcano_plotly}
\alias{volcano_plot_do_label_default}
\alias{volcano_plot_ggplot_call}
\alias{volcano_plot_ly_call}
\title{Volcano Plot}
\usage{
volcano_plot(
  de,
  k,
  labels,
  do.label = volcano_plot_do_label_default,
  ymax = Inf,
  max.overlaps = Inf,
  ggplot_call = volcano_plot_ggplot_call
)

\method{plot}{topic_model_de_analysis}(x, ...)

volcano_plotly(
  de,
  k,
  file,
  labels,
  subsample_below_quantile = 0,
  subsample_rate = 0.1,
  width = 500,
  height = 500,
  title = paste("topic", k),
  plot_ly_call = volcano_plot_ly_call
)

volcano_plot_do_label_default(lfc, z)

volcano_plot_ggplot_call(
  dat,
  y.label,
  topic.label,
  max.overlaps = Inf,
  font.size = 9
)

volcano_plot_ly_call(dat, y.label, title, width, height)
}
\arguments{
\item{de}{An object of class \dQuote{topic_model_de_analysis},
usually an output from \code{\link{de_analysis}}.}

\item{k}{The topic, selected by number or name.}

\item{labels}{Character vector specifying how the points in the
volcano plot are labeled. This should be a character vector with
one entry per LFC estimate (row of \code{de$postmean}). When not
specified, the row names of \code{de$postmean} are used. When
available. labels are added to the plot using
\code{\link[ggrepel]{geom_text_repel}}.}

\item{do.label}{Describe input argument "do.label" here.}

\item{ymax}{Describe input argument "ymax" here.}

\item{max.overlaps}{Argument passed to
\code{\link[ggrepel]{geom_text_repel}}.}

\item{ggplot_call}{The function used to create the plot. Replace
\code{volcano_plot_ggplot_call} with your own function to customize
the appearance of the plot.}

\item{x}{An object of class \dQuote{topic_model_de_analysis},
usually an output from \code{\link{de_analysis}}.}

\item{\dots}{Additional arguments passed to \code{volcano_plot}.}

\item{file}{Save the interactive volcano plot to this HTML
file using \code{\link[htmlwidgets]{saveWidget}}.}

\item{width}{Width of the plot in pixels. Passed as argument
\dQuote{width} to \code{\link[plotly]{plot_ly}}.}

\item{height}{Height of the plot in pixels. Passed as argument
\dQuote{height} to \code{\link[plotly]{plot_ly}}.}

\item{title}{The text used for the plot title.}

\item{plot_ly_call}{The function used to create the plot. Replace
\code{volcano_plot_ly_call} with your own function to customize
the appearance of the interactive plot.}

\item{lfc}{}

\item{z}{}

\item{dat}{A data frame passed as input to
\code{\link[ggplot2]{ggplot}}, containing, at a minimum, columns
\dQuote{label}, \dQuote{mean}, \dQuote{beta}, \dQuote{se},
\dQuote{z}, \dQuote{pval} and \dQuote{label}.}

\item{y.label}{Label to use in the plot for \dQuote{dat$y}.}

\item{topic.label}{The name or number of the topic being plotted.
Only used to determine the plot title.}

\item{font.size}{Font size used in plot.}
}
\value{
A \code{ggplot} object or a \code{plotly} object.
}
\description{
Create a \dQuote{volcano} plot to visualize the
  results of a differential count analysis using a topic model. Here,
  the volcano plot is a scatterplot in which the posterior mean
  log-fold change (LFC), estimated by running the methods implemented
  in \code{\link{de_analysis}}, is plotted against the estimated
  z-score.
}
\details{
The colour of the points is varied by the average count,
on the logarithmic scale; since the evidence (z-score or p-value)
typically increases with more observed counts, the variables with
smallest average counts should usually appear toward the bottom of
the volcano plot. Only points above a specified z-score (or
p-value) quantile are labeled. Note that points with an average
count of zero or less are not shown.

To better accommodate situations in which some z-scores (or
p-values) are much larger than all the others, the z-scores and
-log10 p-values are plotted on the square-root scale. To change
this, as well as other aspects, of the plot, replace
\code{volcano_plot_ggplot_call} with your own function; see input
argument \dQuote{ggplot_call}.

The \dQuote{ggrepel} package is used to arrange the labels in a
visually attractive manner.

Use interactive volcano plot is created using the \dQuote{plotly}
package. The \dQuote{hover text} shows the label (see input
argument \dQuote{labels}) and detailed LFC statistics as they were
calculated by \code{\link{de_analysis}}.
}
\examples{
# See help(de_analysis) for examples.

}
\seealso{
\code{\link{de_analysis}}
}
