% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcano_plot}
\alias{volcano_plot}
\alias{volcano_plot_ggplot_call}
\alias{plot.topic_model_diff_count}
\title{Volcano Plot}
\usage{
volcano_plot(
  diff_count_result,
  k,
  labels,
  y = c("zscore", "pvalue"),
  betamax = 10,
  label_above_quantile = 0.99,
  ggplot_call = volcano_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots)
)

volcano_plot_ggplot_call(dat, y.label, topic.label, font.size = 9)

\method{plot}{topic_model_diff_count}(x, ...)
}
\arguments{
\item{diff_count_result}{An object of class
\dQuote{topic_model_diff_count}, usually an output from
\code{\link{diff_count_analysis}}.}

\item{k}{The topic, or topics, selected by number or name. When not
specified, all topics are plotted.}

\item{labels}{Character vector specifying how the points in the
volcano plot are labeled. This should be a character vector with
one entry per log-fold change statistic (row of
\code{diff_count_result$beta}). When not specified, the row names
of \code{diff_count_result$beta} are used, if available. Labels are
added to the plot using \code{link[ggrepel]{geom_text_repel}}.}

\item{y}{The measure of support to plot in the y-axis: use \code{y
= "zscore"} to plot the z-score magnitudes; use \code{y = "pvalue"}
to plot the -log1 p-values.}

\item{betamax}{Truncate the log-fold change statistics
(\code{beta}) by this amount. Any statistics greater in magnitude
than \code{betamax} are set to \code{betamax} or \code{-betamax}.}

\item{label_above_quantile}{Only z-scores or p-values (depending on
choice of \code{y}) above this quantile are labeled in the volcano
plot. \code{link[ggrepel]{geom_text_repel}} will attempt to label
all points when \code{label_above_quantile = 0}.}

\item{ggplot_call}{The function used to create the plot. Replace
\code{volcano_plot_ggplot_call} with your own function to customize
the appearance of the plot.}

\item{plot_grid_call}{When multiple topics are selected, this is
the function used to arrange the plots into a grid using
\code{\link[cowplot]{plot_grid}}. It should be a function accepting
a single argument, \code{plots}, a list of \code{ggplot} objects.}

\item{dat}{A data frame passed as input to
\code{\link[ggplot2]{ggplot}}, containing, at a minimum, columns
\dQuote{beta}, \dQuote{mean}, \dQuote{y} and \dQuote{label}.}

\item{y.label}{Description of input argument "y.label" goes here.}

\item{topic.label}{The name or number of the topic being plotted.
Only used to determine the plot title.}

\item{font.size}{Font size used in plot.}

\item{x}{Describe input argument "x" here.}
}
\value{
A \code{ggplot} object.
}
\description{
Create one or more "volcano" plots to visualize the
  results of a differential count analysis using a topic model. A
  volcano plot is a scatterplot in which the log-fold change,
  estimated using a multinomial topic model, is plotted against the
  p-value or z-score.
}
\details{
The colour of the points is varied by the average count,
on the logarithmic scale; since the evidence (z-score or p-value)
typically increases with more observed counts, the variables with
smallest average counts should usually appear toward the bottom of
the volcano plot. Only points above a specified z-score (or
p-value) quantile are labeled.

To better accommodate situations in which some z-scores (or
p-values) are much larger than all the others, the z-scores and
-log10 p-values are plotted on the square-root scale. To change
this, as well as other aspects, of the plot, replace
\code{volcano_plot_ggplot_call} with your own function; see input
argument \dQuote{ggplot_call}.

The \dQuote{ggrepel} package is used to arrange the labels in a
visually attractive manner.
}
