% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcano_plot}
\alias{volcano_plot}
\alias{volcano_plot_ggplot_call}
\title{Add Title Here}
\usage{
volcano_plot(diff_count_result, k, ggplot_call = volcano_plot_ggplot_call)

volcano_plot_ggplot_call(dat, font.size = 9)
}
\arguments{
\item{diff_count_result}{Describe input argument "fit" here.}

\item{dat}{Describe input argument "dat" here.}

\item{font.size}{Describe input argument "font.size" here.}
}
\value{
Describe the return value here.
}
\description{
Add description here.
}
\details{
Add details here.
}
